Log axis and normalization refactor
